feat: complete Site Show page redesign with tabbed layout, filters, and workflows
Restructure the site show page into three focused tabs:

Overview tab:
- Floor plan as primary view with floor selector tabs for multi-floor sites
- Zone summary strip with adaptive density (full pills ≤12 zones, compact grid 13+)
- Temperature compliance banner linking to reports
- Alert filters: search + severity + zone (searchable dropdown)
- Create WO from alert dialog, Acknowledge button

Assets tab:
- Device list with search, status filter (All/Online/Offline/Low Battery),
  searchable zone dropdown, sort (Name/Zone/Battery/Status)
- Device rows with model badges, battery bars, last seen
- Add Device dialog with zone assignment (+ new zone creation)
- Work Orders section with + New dialog and View All link

Setup tab:
- Onboarding checklist with clickable items linking to relevant pages
- Configuration dashboard: 9 cards (Alert Rules, Escalation Chains,
  Report Schedules, Verifications, Gateways, Floor Plans, Modules,
  Maintenance Windows, Users) with live counts
- History links (Timeline, Audit Package)
- Site Details card with Edit dialog
- Danger Zone (Suspend) — role-gated

Backend changes:
- Fix time_bucket → standard PostgreSQL (date_trunc + epoch flooring)
- Fix last_seen_at → last_reading_at column reference
- Fix floor plan device filter by floor_id (was showing all devices on every floor)
- Remove alert limit(10) — load all active alerts
- Add configCounts and tempExcursions24h to controller
- Add alert_id to WorkOrder store validation
- Flash created_id on site creation for onboard redirect
- Fix floor_x/floor_y migration: integer → float

Frontend changes:
- Tab state persisted in URL via ?tab= query param
- Role-based visibility (config + danger zone hidden for viewers)
- KPI strip uses real data (gateways, work orders, low battery counts)
- Details bar removed from always-visible (moved to Setup tab)
- Onboarding wizard redesigned with X4 editorial design
- Zone detail page: assign device dialog, alert severity badges
- Sites columns: consistent editorial typography for all headers

New seeder:
- SensorReadingSeeder: 16K+ realistic readings across all devices
- Floor plan SVG with room labels + device placement positions

+++ database/migrations/2026_03_14_100006_create_devices_table.php
@@ -18,8 +18,8 @@ return new class extends Migration
             $table->string('name');
             $table->string('zone')->nullable();
             $table->foreignId('floor_id')->nullable()->constrained('floor_plans')->nullOnDelete();
-            $table->integer('floor_x')->nullable();
-            $table->integer('floor_y')->nullable();
+            $table->float('floor_x')->nullable();
+            $table->float('floor_y')->nullable();
             $table->foreignId('recipe_id')->nullable()->constrained('recipes')->nullOnDelete();
             $table->timestamp('installed_at')->nullable();
             $table->integer('battery_pct')->nullable();
